Make the dev watch work for createPages templates
This fixes the nunjucks engine watch so that it will rebuild the create
pages on build. Without this fix, the dev server explodes with a
template error as all of the context variables are undefined upon watch
fire.

This is probably just a hack that breaks some other functionality but it
makes it so I don't need to kill and restart the dev server every time I
make a change.